Defect fix to check if the user has selected existing line to remove
producst only if added before and engineer visit product is present in
cache.

+++ src/main/java/com/vf/uk/dal/broadband/svc/impl/BroadbandServiceImpl.java
@@ -791,7 +791,7 @@ public class BroadbandServiceImpl implements BroadbandService {
 			AddProductRequest addProductRequest = broadbandJourneyServiceAssembler.addProductRequest(broadband);
 			broadbandDao.updateBasketWithServiceId(addProductRequest, broadband.getBasketId(),
 					broadband.getBasketInfo().getPackageId());
-		} else if ("EXISTING".equalsIgnoreCase(updateLineRequest.getLineTreatmentType()) && StringUtils
+		} else if ("EXISTING".equalsIgnoreCase(updateLineRequest.getLineTreatmentType()) && broadband.getEngineeringVisitCharge() != null && StringUtils
 				.equalsIgnoreCase(serviceIdSku, broadband.getEngineeringVisitCharge().getEngVisitProductId())) {
 			broadbandDao.deleteProductFromBasket(broadband.getBasketId(), broadband.getBasketInfo().getPackageId(),
 					engVisitProductLine);
